refactor(slice-45): extract shared FE test fixture helpers
Move duplicated experiment/vectorDb/explore/detail builders into
frontend/src/test/helpers so suites share Partial<> factories (DECISIONS #158).
