+++ docs/plan/DECISIONS.md
@@ -158,3 +158,4 @@
 | 152 | 2026-07-28 | 45 | Land `server/db/{ports,mongo,postgres}/`; shims at old flat `server.db.*` paths except **no** `server/db/postgres.py` (collides with package); pool API re-exported on `server.db.postgres` package; `schema.sql` co-located with `postgres.py` | Must hotspot #2; preserves Protocols + factory; `from server.db import postgres` / `close_pool` stays valid | Split db across multiple PRs / invent `pg/` package rename |
 | 153 | 2026-07-28 | 45 | Fat-surface slim: `experiments_lifecycle.py` + `cli/display.py`; keep routers/commands as façades | Execution order step 3 after db; pre→post lines experiments 553→388, cli/main 519→246 | Defer slim until FE primitives / leave god modules |
 | 157 | 2026-07-28 | 45 | Split `frontend/src/components/` into `screens/`, `chrome/`, `experiment/`, `stats/` (+ co-located tests); nest `experimentDetail` under `experiment/` | Step 7 after SLAP; App + mocks retargeted; FE floors still green | Leave flat components/ / move without co-located tests |
+| 158 | 2026-07-28 | 45 | FE shared test helpers under `frontend/src/test/helpers/` (`experiments`, `vectorDbStats`, `explore`, `experimentDetail`); screen + stats suites import builders | Step 8 After-Check; `src/test/**` already coverage-excluded; 261 tests / floors unchanged | Leave duplicated builders in mega-suites / invent factories only for new tests |
